Pivotal ID # 172476764: Refresh Interdatabase Support
Fix mapping module name and packages
