+++ commons/commons-model-extended-mapping/src/main/kotlin/ebi/ac/uk/extended/mapping/to/ToAttribute.kt
@@ -1,0 +1,13 @@
+package ebi.ac.uk.extended.mapping.to
+
+import ebi.ac.uk.extended.model.ExtAttribute
+import ebi.ac.uk.extended.model.ExtAttributeDetail
+import ebi.ac.uk.model.Attribute
+import ebi.ac.uk.model.AttributeDetail
+
+internal const val TO_ATTRIBUTE_EXTENSIONS = "ebi.ac.uk.extended.mapping.to.ToAttributeKt"
+
+fun ExtAttribute.toAttribute() = Attribute(name, value, reference, toDetails(nameAttrs), toDetails(valueAttrs))
+
+private fun toDetails(details: List<ExtAttributeDetail>): MutableList<AttributeDetail> =
+    details.mapTo(mutableListOf()) { AttributeDetail(it.name, it.value) }
