+++ commons/commons-model-extended-mapping/src/main/kotlin/ebi/ac/uk/extended/mapping/to/ToTable.kt
@@ -1,0 +1,14 @@
+package ebi.ac.uk.extended.mapping.to
+
+import ebi.ac.uk.extended.model.ExtFileTable
+import ebi.ac.uk.extended.model.ExtLinkTable
+import ebi.ac.uk.extended.model.ExtSectionTable
+import ebi.ac.uk.model.FilesTable
+import ebi.ac.uk.model.LinksTable
+import ebi.ac.uk.model.SectionsTable
+
+internal const val TO_TABLE_EXTENSIONS = "ebi.ac.uk.extended.mapping.to.ToTableKt"
+
+fun ExtSectionTable.toTable(): SectionsTable = SectionsTable(sections.map { section -> section.toSection() })
+fun ExtFileTable.toTable(): FilesTable = FilesTable(files.map { file -> file.toFile() })
+fun ExtLinkTable.toTable(): LinksTable = LinksTable(links.map { link -> link.toLink() })
