+++ submission/persistence/build.gradle.kts
@@ -18,7 +18,7 @@ dependencies {
     api(project(":commons:commons-bio"))
     api(project(":commons:commons-serialization"))
     api(project(":commons:commons-model-extended"))
-    api(project(":commons:commons-model-extended-serialization"))
+    api(project(":commons:commons-model-extended-mapping"))
     api(project(":commons:commons-test"))
 
     implementation(Arrow)
